Ajustement du récit enregistrement pour être compatible avec l'api

+++ app/src/main/java/com/even/domaine/entité/Evenement.kt
@@ -1,11 +1,11 @@
 package com.even.domaine.entité
 
 data class Evenement(
-    val idEvenement: Long,
+    val idEvenement: Int,
     val nomEvenement: String,
     val location: String,
     val date: String,
-    val idOrganisateur: Long,
+    val idOrganisateur: Int,
     val description: String,
     val commentaires: List<Any?>,
     val utilisateurevenements: List<Utilisateurevenement>
